chore(bugfix): fixed parser folder path

+++ src/tests/GRASP_algorithm/GRASP_algorithm_test_capa.cpp
@@ -15,7 +15,7 @@ int main() {
 	miscellaneous::ORLibParser parser;
 	try {
 		Problem problem = parser.parse("../../../src/tests/FicheirosTeste/ORLIB/ORLIB-uncap/a-c/capa.txt");
-        int alpha = 0.1;
+        double alpha = 0.1;
 		algorithm::GRASP grasp(alpha);
 
 		auto solution = grasp.solve(problem);
